authentication_microservices/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
mysqlclient==2.2.0
PyJWT==2.8.0
bcrypt==4.0.1
twilio==8.5.0